requirements.txt
torch
torchvision
pennylane
gradi
numpy
pillow
scikit-learn
